[arcane,utils] Déplace 'ArrayBoundsIndex' dans son propre fichier et unifie les tests de débordement de tableau pour 'ArrayExtents'.
