+++ arcane/src/arcane/utils/srcs.cmake
@@ -119,6 +119,7 @@ set(ARCANE_SOURCES
   VersionInfo.h
   ApplicationInfo.h
   ArcaneGlobal.h
+  ArrayBoundsIndex.h
   ArrayExtents.h
   ArrayBounds.h
   LoopRanges.h
